Refactor MetaLabelSupport
- Lift out of spark-cypher
